moves routes into one resource folder
